Fix lint warnings: unused imports, react-refresh component exports
