+++ src/AddHighlightModel/ProteinToGenomeHighlight.tsx
@@ -1,51 +1,10 @@
 import React from 'react'
 
-import { getSession } from '@jbrowse/core/util'
 import { observer } from 'mobx-react'
 
-import Highlight from './Highlight'
-import {
-  JBrowsePluginProteinStructureModel,
-  JBrowsePluginProteinViewModel,
-} from '../ProteinView/model'
+import ProteinToGenomeHighlightInner from './ProteinToGenomeHighlightInner'
 
-import type { Region } from '@jbrowse/core/util/types'
 import type { LinearGenomeViewModel } from '@jbrowse/plugin-linear-genome-view'
-
-type HighlightField = 'clickGenomeHighlights' | 'hoverGenomeHighlights'
-
-function ProteinToGenomeHighlightInner({
-  model,
-  field,
-}: {
-  model: LinearGenomeViewModel
-  field: HighlightField
-}) {
-  const { assemblyManager, views } = getSession(model)
-  const { assemblyNames } = model
-  const proteinView = views.find(f => f.type === 'ProteinView') as
-    | JBrowsePluginProteinViewModel
-    | undefined
-  const assemblyName = assemblyNames[0]!
-  const assembly = assemblyManager.get(assemblyName)
-  return assembly ? (
-    <>
-      {proteinView?.structures.map(
-        (structure: JBrowsePluginProteinStructureModel, idx: number) =>
-          structure[field].map((r: Region, idx2: number) => (
-            <Highlight
-              key={`${JSON.stringify(r)}-${idx}-${idx2}`}
-              start={r.start}
-              end={r.end}
-              refName={r.refName}
-              assemblyName={assemblyName}
-              model={model}
-            />
-          )),
-      )}
-    </>
-  ) : null
-}
 
 export const ProteinToGenomeClickHighlight = observer(
   function ProteinToGenomeClickHighlight({
@@ -53,7 +12,12 @@ export const ProteinToGenomeClickHighlight = observer(
   }: {
     model: LinearGenomeViewModel
   }) {
-    return <ProteinToGenomeHighlightInner model={model} field="clickGenomeHighlights" />
+    return (
+      <ProteinToGenomeHighlightInner
+        model={model}
+        field="clickGenomeHighlights"
+      />
+    )
   },
 )
 
@@ -63,6 +27,11 @@ export const ProteinToGenomeHoverHighlight = observer(
   }: {
     model: LinearGenomeViewModel
   }) {
-    return <ProteinToGenomeHighlightInner model={model} field="hoverGenomeHighlights" />
+    return (
+      <ProteinToGenomeHighlightInner
+        model={model}
+        field="hoverGenomeHighlights"
+      />
+    )
   },
 )
